+++ components/lv_examples/lv_ex_conf.h
