+++ components/lvgl/CMakeLists.txt
@@ -1,3 +1,3 @@
 file(GLOB_RECURSE SOURCES lvgl/src/*.c)
 idf_component_register(SRCS ${SOURCES}
-                       INCLUDE_DIRS ../ lvgl/ .)
+                       INCLUDE_DIRS . lvgl)
