+++ components/lvgl_ili9341/component.mk
@@ -8,4 +8,4 @@ CFLAGS += -DTOUCH_SUPPORT=1
 
 COMPONENT_SRCDIRS := . 
 	
-COMPONENT_ADD_INCLUDEDIRS := $(COMPONENT_SRCDIRS) . ../ ../lvgl
+COMPONENT_ADD_INCLUDEDIRS := $(COMPONENT_SRCDIRS) .
