+++ main/CMakeLists.txt
@@ -1,4 +1,4 @@
 set(SOURCES main.c)
 idf_component_register(SRCS ${SOURCES}
                     INCLUDE_DIRS .
-                    REQUIRES drv lvgl lv_examples)
+                    REQUIRES lvgl_ili9341 lvgl lv_examples)
